updates to contact method validation. Applicant gender is now a required field
