+++ vsu-app/ClientApp/src/app/shared/components/applicant-information/applicant-information.helper.ts
@@ -1,10 +1,12 @@
 import { FormBuilder, FormGroup, Validators } from "@angular/forms";
 import { ApplicationType, EnumHelper } from "../../enums-list";
 import { POSTAL_CODE } from "../../regex.constants";
+import { ContactMethodHelper } from "../contact-method/contact-method.helper";
 
 export class ApplicantInfoHelper {
     postalRegex = POSTAL_CODE;
     enum = new EnumHelper();
+    contactMethodHelper = new ContactMethodHelper();
     public setupFormGroup(fb: FormBuilder, form_type: ApplicationType): FormGroup {
         let group = {
             applicantType: ['', Validators.required],
@@ -15,6 +17,7 @@ export class ApplicantInfoHelper {
             lastName: ['', Validators.required],
 
             birthDate: ['', [Validators.required]],
+            gender: ['', [Validators.required]],
             genderOther: [''],
 
             preferredLanguage: ['English'],
@@ -29,7 +32,7 @@ export class ApplicantInfoHelper {
             }),
 
             mayWeSendCorrespondence: [this.enum.Boolean.True.val],
-            contactMethods: fb.array([this.createContactMethod(fb, '', true), this.createContactMethod(fb), this.createContactMethod(fb)]),
+            contactMethods: fb.array([this.contactMethodHelper.createContactMethod(fb), this.contactMethodHelper.createContactMethod(fb), this.contactMethodHelper.createContactMethod(fb)]),
             atLeastOneContactMethod: ['', Validators.required],
         }
 
@@ -39,22 +42,13 @@ export class ApplicantInfoHelper {
             group['coveredByVictimServiceProgramComment'] = [''];
 
             group['victimServiceWorker'] = fb.array([]);
-            // group['managerFirstName'] = [''];
-            // group['managerLastName'] = [''];
-            // group['agencyName'] = [''];
-            // group['managerPhone'] = [''];
-            // group['managerEmail'] = ['', [Validators.email]];
 
             group['victimAlreadySubmitted'] = [null];
             group['victimAlreadySubmittedComment'] = [''];
             group['otherFamilyAlsoApplying'] = [null];
             group['otherFamilyAlsoApplyingComment'] = [''];
-            group['gender'] = ['', Validators.required];
             group['supportPersonRelationship'] = [''];
             group['IFMRelationship'] = [''];
-        }
-        else if (form_type === ApplicationType.NOTIFICATION) {
-            group['gender'] = [''];
         }
 
         return fb.group(group);
@@ -72,45 +66,40 @@ export class ApplicantInfoHelper {
         });
     }
 
-    public createContactMethod(fb: FormBuilder, typeString: string = '', required: boolean = false) {
-        let current_validators = [];
-        let typeAndMessageValidators = [];
-        if (required) {
-            current_validators = [Validators.required];
-            typeAndMessageValidators = [Validators.required];
-        }
-        let label = '';
-        let type: number = 0;
-        switch (typeString) {
-            case 'telephone': {
-                type = this.enum.ContactType.Telephone.val;
-                current_validators = current_validators.concat([Validators.minLength(10), Validators.maxLength(15)]);
-                label = 'Telephone Number';
-                break;
-            }
-            case 'mobile': {
-                type = this.enum.ContactType.Cellular.val;
-                current_validators = current_validators.concat([Validators.minLength(10), Validators.maxLength(15)]);
-                label = 'Cellular Number';
-                break;
-            }
-            case 'email': {
-                type = this.enum.ContactType.Email.val;
-                current_validators = current_validators.concat([Validators.email]);
-                label = 'Email Address';
-                break;
-            }
-            default: {
-                type = this.enum.ContactType.Unset.val;
-                break;
-            }
-        }
-        return fb.group({
-            type: [type, typeAndMessageValidators],
-            previousType: [type],
-            val: ['', current_validators],
-            label: [label],
-            leaveMessage: ['', typeAndMessageValidators],
-        });
-    }
+    // public createContactMethod(fb: FormBuilder, typeString: string = '') {
+    //     let current_validators = [];
+    //     let label = '';
+    //     let type: number = 0;
+    //     switch (typeString) {
+    //         case 'telephone': {
+    //             type = this.enum.ContactType.Telephone.val;
+    //             current_validators = current_validators.concat([Validators.minLength(10), Validators.maxLength(15)]);
+    //             label = 'Telephone Number';
+    //             break;
+    //         }
+    //         case 'mobile': {
+    //             type = this.enum.ContactType.Cellular.val;
+    //             current_validators = current_validators.concat([Validators.minLength(10), Validators.maxLength(15)]);
+    //             label = 'Cellular Number';
+    //             break;
+    //         }
+    //         case 'email': {
+    //             type = this.enum.ContactType.Email.val;
+    //             current_validators = current_validators.concat([Validators.email]);
+    //             label = 'Email Address';
+    //             break;
+    //         }
+    //         default: {
+    //             type = this.enum.ContactType.Unset.val;
+    //             break;
+    //         }
+    //     }
+    //     return fb.group({
+    //         type: [type],
+    //         previousType: [type],
+    //         val: ['', current_validators],
+    //         label: [label],
+    //         leaveMessage: [''],
+    //     });
+    // }
 }
