+++ vsu-app/ClientApp/src/app/shared/components/contact-method/contact-method.component.html
@@ -1,8 +1,8 @@
 <div [formGroup]="form" class="contact-methods">
     <div class="row">
         <div class="col-md-4 contact-method-col">
-            <app-field [label]="'Preferred Method of Contact ' + (number + 1)" [required]="isRequired" [valid]="isMyControlValid(form.get('type'))" errorMessage="" [disabled]="disabled">
-                <select class="form-control" formControlName="type" (change)="contactMethodChange(form, isRequired)">
+            <app-field [label]="'Preferred Method of Contact ' + (number + 1)" [required]="false" [valid]="isMyControlValid(form.get('type'))" errorMessage="" [disabled]="disabled">
+                <select class="form-control" formControlName="type" (change)="contactMethodChange(form); checkAtLeastOneContactMethod(parent, isDeligate);">
                     <option [ngValue]="null">Select One</option>
                     <option [value]="enum.ContactType.Telephone.val">Telephone</option>
                     <option [value]="enum.ContactType.Cellular.val">Cellular</option>
@@ -13,7 +13,7 @@
 
         <div class="col-md-4 contact-method-col">
             <ng-container *ngIf="form.get('type').value == enum.ContactType.Telephone.val">
-                <app-field label="Telephone Number" [required]="isRequired" [valid]="isMyControlValid(form.get('val'))" errorMessage="Please enter your phone number" [disabled]="disabled">
+                <app-field label="Telephone Number" [required]="true" [valid]="isMyControlValid(form.get('val'))" errorMessage="Please enter your phone number" [disabled]="disabled">
                     <input class="form-control" mask='000-000-0000' type="text" formControlName="val" *ngIf="parent.get('address.country').value === 'Canada' || parent.get('address.country').value === 'United States of America'; else phone_no_mask" (change)="checkAtLeastOneContactMethod(parent, isDeligate)">
                     <ng-template #phone_no_mask>
                         <input class="form-control" type="text" formControlName="val" digitOnly (change)="checkAtLeastOneContactMethod(parent, isDeligate)">
@@ -22,7 +22,7 @@
             </ng-container>
 
             <ng-container *ngIf="form.get('type').value == enum.ContactType.Cellular.val">
-                <app-field label="Cellular Number" [required]="isRequired" [valid]="isMyControlValid(form.get('val'))" errorMessage="Please enter your cell number" [disabled]="disabled">
+                <app-field label="Cellular Number" [required]="true" [valid]="isMyControlValid(form.get('val'))" errorMessage="Please enter your cell number" [disabled]="disabled">
                     <input class="form-control" mask='000-000-0000' type="text" formControlName="val" *ngIf="parent.get('address.country').value === 'Canada' || parent.get('address.country').value === 'United States of America'; else cell_no_mask" (change)="checkAtLeastOneContactMethod(parent, isDeligate)">
                     <ng-template #cell_no_mask>
                         <input class="form-control" type="text" formControlName="val" digitOnly (change)="checkAtLeastOneContactMethod(parent, isDeligate)">
@@ -31,14 +31,14 @@
             </ng-container>
 
             <ng-container *ngIf="form.get('type').value == enum.ContactType.Email.val">
-                <app-field label="Email Address" [required]="isRequired" [valid]="isMyControlValid(form.get('val'))" errorMessage="Please enter your email address" [disabled]="disabled">
+                <app-field label="Email Address" [required]="true" [valid]="isMyControlValid(form.get('val'))" errorMessage="Please enter your email address" [disabled]="disabled">
                     <input class="form-control" type="text" formControlName="val" maxlength="100" (change)="checkAtLeastOneContactMethod(parent, isDeligate)">
                 </app-field>
             </ng-container>
         </div>
 
         <div class="col-md-4 contact-method-col" *ngIf="form.get('type').value != enum.ContactType.Unset.val">
-            <app-field label="May we leave a detailed message?" errorMessage="" [disabled]="disabled" [required]="isRequired" [valid]="isMyControlValid(form.get('leaveMessage'))" errorMessage="Please choose an option">
+            <app-field label="May we leave a detailed message?" errorMessage="" [disabled]="disabled" [required]="true" [valid]="isMyControlValid(form.get('leaveMessage'))" errorMessage="Please choose an option">
                 <label class="inline-label">
                     <input type="radio" [value]="enum.Boolean.True.val" formControlName="leaveMessage" (change)="checkAtLeastOneContactMethod(parent, isDeligate)">
                     <span>Yes</span>
